fix: consolidate API lifecycle cleanup helpers for pkg-file-count
Move runtime session drain and project-state removal helpers into server_durable_session_read_test.go so pkg/api/servertests stays within the 15-file package limit.
